[FIX] 폴더구조 변경으로 인한 merge 진행
