+++ src/main/java/nova/mjs/admin/account/repository/AdminRepository.java
@@ -1,10 +1,10 @@
 package nova.mjs.admin.account.repository;
 
-import nova.mjs.admin.account.entity.Admin;
+import nova.mjs.admin.account.entity.StudentCouncilAdmin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AdminRepository extends JpaRepository<Admin, Long> {
-    Optional<Admin> findByAdminId(String adminId);
+public interface AdminRepository extends JpaRepository<StudentCouncilAdmin, Long> {
+    Optional<StudentCouncilAdmin> findByContactEmail(String adminId);
 }
