+++ src/main/java/nova/mjs/admin/department_schedule/service/AdminDepartmentScheduleService.java
@@ -3,14 +3,14 @@ package nova.mjs.admin.department_schedule.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nova.mjs.admin.account.entity.Admin;
+import nova.mjs.admin.account.entity.StudentCouncilAdmin;
 import nova.mjs.admin.account.exception.AdminIdMismatchException;
 import nova.mjs.admin.account.repository.AdminRepository;
 import nova.mjs.admin.department_schedule.dto.AdminDepartmentScheduleRequestDTO;
 import nova.mjs.admin.department_schedule.dto.AdminDepartmentScheduleResponseDTO;
 import nova.mjs.admin.department_schedule.repository.AdminDepartmentScheduleRepository;
-import nova.mjs.department.entity.Department;
-import nova.mjs.department.entity.DepartmentSchedule;
+import nova.mjs.domain.department.entity.Department;
+import nova.mjs.domain.department.entity.DepartmentSchedule;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -28,7 +28,7 @@ public class AdminDepartmentScheduleService {
     @Transactional
     public AdminDepartmentScheduleResponseDTO create(String adminId, AdminDepartmentScheduleRequestDTO request) {
 
-        Admin admin = adminRepository.findByAdminId(adminId)
+        StudentCouncilAdmin studentCouncilAdmin = adminRepository.findByContactEmail(adminId)
                 .orElseThrow(() -> {
                     log.warn("[일정 생성 실패] 존재하지 않는 adminId: {}", adminId);
                     return new AdminIdMismatchException();
@@ -40,7 +40,7 @@ public class AdminDepartmentScheduleService {
                 .colorCode(request.getColorCode())
                 .startDate(request.getStartDate())
                 .endDate(request.getEndDate())
-                .department(admin.getDepartment())
+                .department(studentCouncilAdmin.getDepartment())
                 .departmentScheduleUuid(UUID.randomUUID())
                 .build();
 
@@ -51,10 +51,10 @@ public class AdminDepartmentScheduleService {
     }
 
     public List<AdminDepartmentScheduleResponseDTO> getSchedulesByMonth(String adminId, int year, int month) {
-        Admin admin = adminRepository.findByAdminId(adminId)
+        StudentCouncilAdmin studentCouncilAdmin = adminRepository.findByContactEmail(adminId)
                 .orElseThrow(AdminIdMismatchException::new);
 
-        Department department = admin.getDepartment();
+        Department department = studentCouncilAdmin.getDepartment();
         LocalDate start = LocalDate.of(year, month, 1);
         LocalDate end = start.withDayOfMonth(start.lengthOfMonth());
