+++ src/main/java/nova/mjs/domain/department/entity/Department.java
@@ -1,8 +1,9 @@
-package nova.mjs.department.entity;
+package nova.mjs.domain.department.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import nova.mjs.member.entity.enumList.College;
+import nova.mjs.domain.member.entity.enumList.College;
+import nova.mjs.domain.member.entity.enumList.DepartmentName;
 import nova.mjs.util.entity.BaseEntity;
 
 import java.util.ArrayList;
@@ -13,8 +14,8 @@ import java.util.UUID;
 @Getter
 @Builder
 @AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "departments")
+@NoArgsConstructor
+@Table(name = "department")
 public class Department extends BaseEntity {
 
     @Id
@@ -25,11 +26,12 @@ public class Department extends BaseEntity {
     @Column(nullable = false, unique = true)
     private UUID departmentUuid;
 
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private String departmentName;
+    private DepartmentName departmentName;
 
     @Column
-    private String studentCouncilName;
+    private String studentCouncilName; // 학생회 명
 
     @Column
     private String studentCouncilLogo;
@@ -50,13 +52,14 @@ public class Department extends BaseEntity {
     @Column(nullable = false)
     private College college;
 
+    @Builder.Default
     @OneToMany(mappedBy = "department", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<DepartmentSchedule> schedules = new ArrayList<>();
-
+    @Builder.Default
     @OneToMany(mappedBy = "department", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<DepartmentNotice> notices = new ArrayList<>();
 
-    public void updateInfo(String departmentName, String studentCouncilName, String homepageUrl, String instagramUrl, String description, String studentCouncilLogo, String slogan, College college) {
+    public void updateInfo(DepartmentName departmentName, String studentCouncilName, String homepageUrl, String instagramUrl, String description, String studentCouncilLogo, String slogan, College college) {
         this.departmentName = departmentName;
         this.studentCouncilName = studentCouncilName;
         this.homepageUrl = homepageUrl;
