+++ src/main/java/nova/mjs/domain/department/entity/DepartmentNotice.java
@@ -1,10 +1,9 @@
-package nova.mjs.department.entity;
+package nova.mjs.domain.department.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 import nova.mjs.util.entity.BaseEntity;
 
-import java.time.LocalDate;
 import java.util.UUID;
 
 @Entity
@@ -12,32 +11,25 @@ import java.util.UUID;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "department_schedules")
-public class DepartmentSchedule extends BaseEntity {
+@Table(name = "department_notices")
+public class DepartmentNotice extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false, unique = true)
-    private UUID departmentScheduleUuid;
+    private UUID uuid;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "department_id")
     private Department department;
 
-    @Column(nullable = false, length = 20)
+
+    @Column(nullable = false)
     private String title;
-
-    @Column(nullable = false)
-    private String colorCode;
-
-    @Column(nullable = false)
-    private LocalDate startDate;
-
-    @Column(nullable = false)
-    private LocalDate endDate;
 
     @Column
     private String content;
+
 }
