+++ src/main/java/nova/mjs/domain/department/entity/DepartmentSchedule.java
@@ -1,11 +1,10 @@
-package nova.mjs.department.entity;
+package nova.mjs.domain.department.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 import nova.mjs.util.entity.BaseEntity;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.time.LocalDate;
 import java.util.UUID;
 
 @Entity
@@ -13,24 +12,32 @@ import java.util.UUID;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "department_notices")
-public class DepartmentNotice extends BaseEntity {
+@Table(name = "department_schedules")
+public class DepartmentSchedule extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false, unique = true)
-    private UUID uuid = UUID.randomUUID();
+    private UUID departmentScheduleUuid;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "department_id")
     private Department department;
 
-    @Column(nullable = false)
+    @Column(nullable = false, length = 20)
     private String title;
+
+    @Column(nullable = false)
+    private String colorCode;
+
+    @Column(nullable = false)
+    private LocalDate startDate;
+
+    @Column(nullable = false)
+    private LocalDate endDate;
 
     @Column
     private String content;
-
 }
