+++ src/main/java/nova/mjs/domain/department/service/DepartmentScheduleService.java
@@ -1,13 +1,13 @@
-package nova.mjs.department.service;
+package nova.mjs.domain.department.service;
 
 
 import lombok.RequiredArgsConstructor;
-import nova.mjs.department.DTO.DepartmentScheduleResponseDTO;
-import nova.mjs.department.entity.Department;
-import nova.mjs.department.entity.DepartmentSchedule;
-import nova.mjs.department.exception.DepartmentNotFoundException;
-import nova.mjs.department.repository.DepartmentRepository;
-import nova.mjs.department.repository.DepartmentScheduleRepository;
+import nova.mjs.domain.department.DTO.DepartmentScheduleResponseDTO;
+import nova.mjs.domain.department.entity.Department;
+import nova.mjs.domain.department.entity.DepartmentSchedule;
+import nova.mjs.domain.department.exception.DepartmentNotFoundException;
+import nova.mjs.domain.department.repository.DepartmentRepository;
+import nova.mjs.domain.department.repository.DepartmentScheduleRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
